[kernel] E: Move ScoreBoard to Its Own Module
